refactor: Follow PEP8 standards in naming file names of models
