+++ exam_django/asset/models/__init__.py
@@ -1,10 +1,10 @@
-from .Asset import Asset
-from .AssetType import AssetType
-from .BusinessUnit import BusinessUnit
-from .Employee import Employee
-from .Location import Location
-from .Memory import Memory
-from .AssetLog import AssetLog
+from .asset import Asset
+from .asset_type import AssetType
+from .business_unit import BusinessUnit
+from .employee import Employee
+from .location import Location
+from .memory import Memory
+from .asset_log import AssetLog
 
 __all__ = [
     "Asset",
