+++ exam_django/asset/service/asset_assign_crud_service/assign_asset_service.py
@@ -13,7 +13,7 @@ from messages import (
     ASSET_NOT_FOUND,
 )
 from asset.models import Employee
-from notification.service.EmailService import EmailService
+from notification.service.email_service import EmailService
 from rest_framework.renderers import JSONRenderer
 from asset.serializers import AssignAssetSerializer
 from asset.models import Asset
